Remove camera event
… and move all camera types from events to commands.
