+++ commands/Cargo.toml
@@ -15,4 +15,5 @@ url = { version = "2.1.1", features = ["serde"] }
 uuid = { version = "0.8", features = ["serde"] }
 
 [dev-dependencies]
+serde_json = "1.0"
 uuid = { version = "0.8", features = ["v4", "serde"] }
