+++ events/Cargo.toml
@@ -12,6 +12,3 @@ serde = { version = "1.0", features = ["derive"] }
 url = { version = "2.1.1", features = ["serde"] }
 uuid = { version = "0.8", features = ["serde"] }
 zvariant = { version = "2.4.0", default-features = false }
-
-[dev-dependencies]
-serde_json = "1.0"
